Split GUIControl into a base class ClassicSimConrol and a subclass GUIControl
